refactoring on all screens
