+++ UmbrellaSharing-iOS/Layouts/Base.lproj/Main.storyboard
@@ -189,9 +189,6 @@
                         </constraints>
                         <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
                     </view>
-                    <connections>
-                        <outlet property="informationButton" destination="vMD-MA-ca3" id="Cba-ed-7q4"/>
-                    </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
             </objects>
