+++ UmbrellaSharing-iOS/ViewControllers/QRCodeScreenViewController.swift
@@ -10,28 +10,29 @@ import Foundation
 import UIKit
 import Toast_Swift
 
-class QRCodeScreenViewController: UIViewController, QRDataModelDelegate {
-
-    @IBOutlet weak var continueButton: UmbrellaButton!
-    @IBOutlet weak var backButton: UmbrellaButton!
+class QRCodeScreenViewController: UIViewController {
+    
+    // MARK: Outlets
+    
     @IBOutlet weak var qrCodeImageView: UIImageView!
     
-    let qrViewModel = QRViewModel()
+    // MARK: Public
     
     var operationType: UmbrellaUtil.OperationType?
     var orderInformation: OrderInformation = OrderInformation()
     
+    // MARK: Private
+    private let qrViewModel = QRViewModel()
+    
+    // MARK: Initialization
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         qrViewModel.delegate = self
-        initView()
-    }
-    
-    private func initView() {
-        continueButton.setTitle("Continue", for: .normal)
-        backButton.setTitle("Go Back", for: .normal)
         presentQR()
     }
+    
+    // MARK: Private methods
     
     private func presentQR() {
         self.view.makeToastActivity(.center)
@@ -47,36 +48,6 @@ class QRCodeScreenViewController: UIViewController, QRDataModelDelegate {
                 qrViewModel.getQRCode(userId: userId, isBuy: isBuy)
             }
         }
-    }
-    
-    @IBAction func pressContinue(_ sender: Any) {
-        if let orderId = orderInformation.orderId, let operationType = operationType {
-            if let code = orderInformation.code {
-                // TODO: Level 3 - Remove after complete implementation
-                // Insert this request in postman:
-                print("https://us.2ssupport.ru/order/getOrderInfo?orderId=\(orderId)&code=\(code)&qrType=\(operationType.rawValue)")
-            }
-            qrViewModel.canWeProceed(orderId: orderId, qrType: operationType)
-        }
-    }
-    
-    @IBAction func back(_ sender: Any) {
-        dismiss(animated: true, completion: nil)
-    }
-    
-    // TODO: Level 3 - Understand what does it mean
-    func generateQRCode(from string: String) -> UIImage? {
-        let data = string.data(using: String.Encoding.ascii)
-
-        if let filter = CIFilter(name: "CIQRCodeGenerator") {
-            filter.setValue(data, forKey: "inputMessage")
-            let transform = CGAffineTransform(scaleX: 3, y: 3)
-
-            if let output = filter.outputImage?.transformed(by: transform) {
-                return UIImage(ciImage: output)
-            }
-        }
-        return nil
     }
     
     private func openHomeScreen() {
@@ -102,17 +73,50 @@ class QRCodeScreenViewController: UIViewController, QRDataModelDelegate {
         self.present(newViewController, animated: true, completion: nil)
     }
     
+    private func rememberThatRentHasBeenStarted(orderId: Int, rentStartDate: Date) {
+        let informationAboutLastSession = InformationAboutLastSession(hasRentStarted: true, orderId: orderId, rentStartDate: rentStartDate)
+        if let informationAboutLastSession = informationAboutLastSession {
+            GlobalDataStorage.shared.saveInformationAboutLastSession(informationAboutLastSession)
+        }
+    }
+    
+    // MARK: IB Actions
+    
+    @IBAction func pressContinue(_ sender: Any) {
+        if let orderId = orderInformation.orderId, let operationType = operationType {
+            if let code = orderInformation.code {
+                // TODO: Level 4 - Remove after complete implementation
+                // Insert this request in postman:
+                print("https://us.2ssupport.ru/order/getOrderInfo?orderId=\(orderId)&code=\(code)&qrType=\(operationType.rawValue)")
+            }
+            qrViewModel.canWeProceed(orderId: orderId, qrType: operationType)
+        }
+    }
+    
+    @IBAction func back(_ sender: Any) {
+        dismiss(animated: true, completion: nil)
+    }
+    
+    @IBAction func openInformation(_ sender: Any) {
+        let storyBoard: UIStoryboard = UIStoryboard(name: "InformationScreen", bundle: nil)
+        let newViewController = storyBoard.instantiateViewController(withIdentifier: "InformationScreenViewController") as! InformationScreenViewController
+        newViewController.modalPresentationStyle = .fullScreen
+        self.present(newViewController, animated: true, completion: nil)
+    }
+}
+
+extension QRCodeScreenViewController: QRDataModelDelegate {
     func didLoadQRCode(orderId: Int, code: Int) {
         self.view.hideToastActivity()
         orderInformation.orderId = orderId
         orderInformation.code = code
-        qrCodeImageView.image = generateQRCode(from: String(code))
+        qrCodeImageView.image = qrViewModel.generateQRCode(from: String(code))
     }
     
     func didLoadReturnCode(code: Int) {
         self.view.hideToastActivity()
         self.orderInformation.code = code
-        qrCodeImageView.image = generateQRCode(from: String(code))
+        qrCodeImageView.image = qrViewModel.generateQRCode(from: String(code))
     }
     
     func qrCodeHasBeenScanned(startTime: Date) {
@@ -142,17 +146,4 @@ class QRCodeScreenViewController: UIViewController, QRDataModelDelegate {
         self.view.makeToast("The operator must scan the code first", duration: 2.0, position: .bottom)
     }
     
-    private func rememberThatRentHasBeenStarted(orderId: Int, rentStartDate: Date) {
-        let informationAboutLastSession = InformationAboutLastSession(hasRentStarted: true, orderId: orderId, rentStartDate: rentStartDate)
-        if let informationAboutLastSession = informationAboutLastSession {
-            GlobalDataStorage.shared.saveInformationAboutLastSession(informationAboutLastSession)
-        }
-    }
-    
-    @IBAction func openInformation(_ sender: Any) {
-        let storyBoard: UIStoryboard = UIStoryboard(name: "InformationScreen", bundle: nil)
-        let newViewController = storyBoard.instantiateViewController(withIdentifier: "InformationScreenViewController") as! InformationScreenViewController
-        newViewController.modalPresentationStyle = .fullScreen
-        self.present(newViewController, animated: true, completion: nil)
-    }
 }
